fixed problem with NumberFormatException

+++ app/src/main/java/ru/msokolov/cryptocurrencyapp/adapter/entities/Currency.kt
@@ -12,7 +12,7 @@ data class Currency(
     val priceChangePercentage24h: Float,
     val marketCapChangePercentage24h: Float,
     val circulatingSupply: Double,
-    val totalSupply: Long,
+    val totalSupply: Float,
     val ath: Float,
     val athChangePercentage: Float
 )
